significant restructure + change all imports + change to match new python library versions
constants - contains all constants (days, locations, styling)
core - all common/shared functionality (time functions, excel writing, styling, xml parsing)
converters - file converters (hastus/ITOPS)
first_last - all firstlast stuff + new graph
gui - all visual stuff / pop up boxes etc
plans - ngr planning code
reports - all excel based reports (error checker, runinfo, tripcount)
stabling - stabling stuff
rsx - all files that change or output rsx (e.g sectoriseRSX, slicer)
run renamer - all run renamer stuff, havent really touched this yet
timetables - all timetable outputs (public/working timetable, timetable summary)
