+++ src/taipan/core/utils.py
