+++ src/taipan/core/xml_parser.py
