+++ src/taipan/core/xml_processor.py
@@ -3,10 +3,9 @@ from typing import List, Dict
 from taipan.constants.trains import SORT_ORDER_UNIT
 from taipan.constants.days import WEEKDAY_KEYS_MASTER, ID_TO_LONG
 
-from taipan.utils import timetrim
-from taipan.xml_parser import resolve_DoO
-from taipan.utils import csl
-from taipan.utils import _time_key
+from taipan.core.utils import timetrim
+from taipan.core.xml_parser import resolve_DoO
+from taipan.core.utils import csl, _time_key
 import numpy as np
 
 def init_store(locations_dict, day_codes):
