+++ src/taipan/plans/NGRDailyPlan.py
