+++ src/taipan/plans/NGRWeeklyPlan.py
