+++ src/taipan/reports/TripCount.py
